requirements.txt
streamlit
pandas
numpy
websocket-client
statsmodels
sqlalchemy
requests
pykalman
scikit-learn